requirements.txt
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.4.0
Pillow>=9.0.0
requests>=2.25.0 